requirements.txt
Django==1.7.5
MySQL-python==1.2.5
Pillow==2.7.0
argparse==1.2.1
djangorestframework==3.1.1
wsgiref==0.1.2
